small code warning fix
git-svn-id: http://svn.miranda-ng.org/main/trunk@6521 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/AVS/src/options.cpp
@@ -297,7 +297,7 @@ static INT_PTR CALLBACK DlgProcOptionsProtos(HWND hwndDlg, UINT msg, WPARAM wPar
 		{
 			LVITEM item = {0};
 			LVCOLUMN lvc = {0};
-			UINT64 newItem = 0;
+			int newItem = 0;
 
 			dialoginit = TRUE;
 			TranslateDialogDefault(hwndDlg);
